Refactored collectible collision
