+++ Assets/Prefabs/Characters.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 8570290824633460431}
   - component: {fileID: 8570290824633460430}
   - component: {fileID: 8570290824633460424}
+  - component: {fileID: 6933011211836863365}
   m_Layer: 8
   m_Name: Player
   m_TagString: Player
@@ -121,6 +122,18 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   speed: 0.13
   cam: {fileID: 0}
+--- !u!114 &6933011211836863365
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8570290824633460429}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 059804e1e7ac8934ba7ffc7f11f7a0de, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &8570290825126533332
 GameObject:
   m_ObjectHideFlags: 0
@@ -133,6 +146,7 @@ GameObject:
   - component: {fileID: 8570290825126533334}
   - component: {fileID: 8570290825126533328}
   - component: {fileID: 8570290825126533331}
+  - component: {fileID: 8614389766345449674}
   m_Layer: 8
   m_Name: Player2
   m_TagString: Player
@@ -242,6 +256,18 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   speed: 0.13
   cam: {fileID: 0}
+--- !u!114 &8614389766345449674
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8570290825126533332}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 059804e1e7ac8934ba7ffc7f11f7a0de, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &8570290826083046101
 GameObject:
   m_ObjectHideFlags: 0
