+++ Assets/Scripts/Environment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 061ab253509700749a0538fafa12c4fa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
