+++ Assets/Scripts/Environment/Collideable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6e553b12e9887724bae3ae7663a9836e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
