+++ Assets/Scripts/Player/CheckCollectibles.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 059804e1e7ac8934ba7ffc7f11f7a0de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
